Revise nav, set files-read-status message.
Revise navigation to store the activity in the state. In the activity's
onCreate, it dispatches an action to set itself as the state's activity;
in onDestroy, it dispatches an action to clear the state's activity.

The state's canReadInstructionsFileMessage file-reading status message
was not being set in the InstructionFiles reducer. I fixed that.