Fix navigation - initial concept was wrong.
Anvil re-renders by calling a view's view method, NOT by (logically)
destroying and re-rendering it. The data that the view method references
must be current data from the store. Therefore, top-level views, in their
view methods, must pass a subtree of store.state to child views.

Also, the nagivation stack does not save props with which views were
constructed. 1. I'm not constructing top-level views (i.e., the things
to which we navigate) with props. 2. I want views to have the current
data in the store and not the data with which they were constructed.

My initial idea for n