Nav: root view dispatches on scene to view fns. Clean up.
There is no need to setContentView of the activity to different root
views. Rather, setContentView to a single root view and in it call
various view functions based on the scene currently at the top of the
navigation stack.
Note that this means that the state does not need to hold a reference to
the activity on which to call setContentView. And we don't need actions
to set activity, clear activity, or show current view as a result.

I converted scenes from strings to a sealed-class faux union-type.

TODO: Since the compiler does not check that the when on scenes in the
root view or the when on actions in the reducer are exhaustive,
respectively, I'm temporarily throwing an error in the else.
I WOULD LIKE A COMPILE-TIME CHECK. Failing that, you probably don't just
want to let the application die - how about submitting an error report
and returning the main screen?

I overrode toString for several classes in order to make the debugging
output easier to read. It would be nice to use a JSON-serialization
library for the values of certain properties. I'm getting errors on
trying to include Jackson, though, and it isn't that big a deal right
now.