Force compiler to check that nav "when" is exhaustive.
As of now (Kotlin 1.0.3), the compiler only checks that a when is
exhaustive if the result of the when expression is used. Since views are
void methods, the result is not used. In order to force the check for
exhaustiveness, I bound the result of the when to a value that is never
used.

Note that I can't do this for the when on action types in the reducer
because the type of the action, as per the way Bansa is written, is
com.brianegan.bansa.Action, which is an interface and not a Kotlin
sealed class.
com.cws.image.Action implements com.brianegan.bansa.Action but the when
is not on a com.cws.image.Action.