Collect and display instruction-parsing failures.
I let the Rx Observables handle failures in the instruction-files
pipeline instead of using Results. (Yes, I'm using exceptions instead of
values with this change but this IS Java-world, after all.)

Failing to parse a malformatted file name is not a failure in the
application but a result that needs to be reported and acted upon. At
this point, that means displaying the failures on the screen so that
whoever copied the files to the device may rename the relevant files.