Coarsely handle MediaPlayer errors.
In the PlayInstructionSequence middleware, I defined a
MediaPlayer.OnErrorListener as a lambda that dispatches
Action.CouldNotPlayInstruction on any error.

Note that the values of the what and extra parameters provide
information about why an error occcured and that some errors are
recoverable merely by starting over (new MediaPlayer, prepare, start).

This coarse handling will do for now, though.