Re-add playing instruction audio.
This mix of app-level state and work with multiple activities is no
good:

1. When an instruction finishes playing, I send a message from the view
model, which holds the media player (should not be affected by config
changes). Both the PlayInstructionActivity and the MainActivity receive
the message: the PlayInstructionActivity has evidently not had onDestroy
called (where it unsubscribes from the view model's messages) before the
MainActivity subscribes.

2. I want to display a snackbar with a message if playing an instruction
errors out. Since such a failure finishes the PlayInstructionActivity
and returns to the MainActivity, I want to show the snackbar in the
MainActivity. I don't know that the MainActivity will be subscribed to
the view model's messages in time to receive the error message.
Therefore, I used startActivityForResult and onActivityResult, going
around the view model's messaging channel. I.e., my method isn't
working.

Either use a single activity or go full normal-Android with worker
fragments and services and stuff.