Fix audio playback to continue thru config change.
I was releasing the media player and setting that property of the view
model to null in PlayInstructionActivity#onDestroy. That method is
called on configuration changes and calling relase while the media
player was playing sent it to an error state. The result was that a
config change stopped the audio and displayed an error message in the
main screen snackbar.

I corrected the behavior to continue playing audio through configuration
changes. Pressing back still stops the audio appropriately.