Get previously-parsed instructions if not dirty.
If ViewModel#needToRefreshInstructions is false, then return the
previously-parsed instructions from ViewModel#getInstructions.

Returning something when ViewModel#needToRefreshInstructions is false,
regardless of how that something was obtained, allows the language tabs,
instructions, and unparsable instructions to be restored on activity
re-creation, such as on a configuration change.

Also, transforming the parsed instructions, setting the view model
properties, preparing the message payload and sending the message is now
done on the Schedulers.io thread. (Schedulers.computation might make
more sense :/  )