Ensure AudioManger is obtained.
I switched an unsafe cast of a MAYBE-audio-service to AudioManger to an
unsafe cast of an audio-service (not maybe) to AudioManger.

Donn Felker says in
http://www.donnfelker.com/why-i-use-crashlytics-part-2/ that 50% - 80%
of errors will initially be null pointer exceptions from fragments
calling getActivity and assuming an activity, not null, will be
returned.

Therefore, I used context.getSystemService instead of
activity?.getSystemService.