Add user feedback for kiosk mode stuff. Log errors.
* User feedback on setting password.
* Error messages in dialogs.
* User feedback and error logging on actions that shouldn't be doable in
  the given state.
* Log errors that shouldn't happen, like failing to write a password to
  shared preferences.

NOTE that if the user enters a password to exit kiosk mode but no
password can be retrieved, then I exit lock-task mode, unlocking the
device from running the app. THIS IS PROBABLY A SECURITY HOLE. I'd
rather unlock the device than make it (at least for me) difficult or
impossible to recover. And note that:
* It should not be possible to enter kiosk mode unless a password has
  been set. Setting means that the password was successfully written.
* The password is currently written to private shared preferences, which
  should only be accessible by this app (and any other with the same
  ID). It seems like this is as good as it gets on Android at this
  point.
* The device is not to have Developer Options enabled while deployed so
  it SHOULDN'T be possible to tamper with the private shared preferences
  via ADB.

You can always change this if we/customers want a different trade-off.