Simplify CueTimer events, draw mutable objects.
From the outside, I want to know whether the timer has initialized. I
can figure out whether it should have completed or not. (It may get
slightly off as it's recreated on configuration changes.)
Therefore, I added a hasInitialized observable and got rid of
hasCompleted.

All the recommendations I've read say not to initialize objects in
onDraw. Since the CueTimer was starting jerky in my app (not the testbed
activity), I changed all of the nice immutable points to mutable points.
It didn't help but drawing the view may be a little faster now :/