Add CueTimer to PlayInstructionActivity.
The CueTimer kept jerking from the initial position to about the first
half second of its rotation. THIS ONLY HAPPENS WHEN THE TIMER IS
STARTED, not when the view is recreated on configuration change.

The Stack Overflow recommendations I found did not help
* Waiting for Activity#onWindowFocusChanged
* Waiting for CueTimer#onAttachedToWindow, though I left that in

I hate to do it but I added a 250-millisecond delay to starting the
instruction. (Only when it is started, not on recreation.)