Make CueTimer-related observables act as expected.
I happened to change the cue duration of an instruction (I renamed the
file) while the app was running. Tapping the button of the
now-non-existent instruction crashed the app instead of causing a
snackbar to display an error message.

I had a when on the value of a BehaviorSubject to make the view model
for PlayInstructionActivity. When an error was thrown in the observable,
the when expression was getting something that was none of the cases and
a missing-case exception (I forget the class) was thrown.
NOTE: This seems like a compiler bug (1.1.2, currently) because that
code should not have compiled.

Instead of putting the when in a try-catch, I make and set the view
model based on the first InstructionEvent, in PlayInstructionPresenter.
NOTE: Without the take(1), this would set a new view model on every
InstructionEvent.

That led to logging emissions from observables, at which point I found
that data binding seems to initially call binding adapter functions with
zero-values for the respective data types and then call the binding
adapters with the data-bound values.
I therefore filtered the timer durations for positive values to avoid
initializing the CueTimer with the zero-values.

Finally, I cleaned up the synchronization streams in CueTimer:
* Timer duration and elapsed time are now combined with zip instead of
  combineLatest.
* The clock-face center and arc radius are set together; they are now
  one emission to one stream instead of an emission to each of two
  streams that were then combineLatest'd, which was causing two
  initializations.