Show rationale for storage permission when needed.
I used a Snackbar to explain why the permission is needed.
I tried to use a DialogFragment at first but I kept getting an fatal
error if I denied the permission, which caused the rationale dialog to
be shown a second time.

The error was
  java.lang.IllegalStateException: Can not perform this action after onSaveInstanceState
This article explains it:
  http://www.androiddesignpatterns.com/2013/08/fragment-transaction-commit-state-loss.html
Since I was using a DialogFragment, though, I was never using commit so
I didn't see where to use commitAllowingStateLoss as a workaround (since
the dialog had no state anyway).

How do these libraries do it? Take a look when you have time:
* https://github.com/hotchemi/PermissionsDispatcher
* https://github.com/Karumi/Dexter
* https://github.com/KingsMentor/Rationale